backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.7
pypdf2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1